requirements.txt
asyncpg>=0.29.0
sqlalchemy[asyncio]>=2.0
pgvector>=0.3.0
pydantic>=2.0
jieba>=0.42.1
fastapi>=0.110.0
uvicorn>=0.29.0
numpy>=1.24.0
cachetools>=5.0
lightrag-hku
pgvector
asyncpg